Add linguist override for .vb files
I noticed that you use the .vb extension for some of your VBA code. However, this extension is marked by Linguist as a VB.NET extension.
To fix that, I've added an override, so the repo can be properly marked as a VBA repo (in the language breakdown and in searches).